fix bool conversions in validation

+++ power_grid_model_c/power_grid_model/include/power_grid_model/exception.hpp
@@ -150,7 +150,7 @@ class InvalidShortCircuitType : public PowerGridError {
     }
     InvalidShortCircuitType(bool sym, FaultType short_circuit_type) {
         append_msg("The short circuit type (" + std::to_string(static_cast<IntS>(short_circuit_type)) +
-                   ") does not match the calculation type (symmetric=" + std::to_string(sym) + ")\n");
+                   ") does not match the calculation type (symmetric=" + std::to_string(static_cast<int>(sym)) + ")\n");
     }
 };
